refactor: rename methods, variables and other minor code improvement

+++ src/verification/verification.service.ts
@@ -9,41 +9,51 @@ import { errorMessage } from '@/common/messages';
 
 @Injectable()
 export class VerificationService {
-	private readonly minutesToMakeNewOtpRequest = 2;
-	private readonly otpExpirationMinutes = 15;
+	private readonly minutesToMakeNewOtpRequest = 2 * 60 * 1000;
+	private readonly otpExpirationMinutes = 15 * 60 * 1000;
 	constructor(
 		@InjectRepository(Verification)
 		private otpRepository: Repository<Verification>,
 	) {}
-
-	async createOtp(user_id: string): Promise<string> {
+	/**
+	 * Creates a new record in database with otp code for a particular user id and returns it.
+	 * This first deletes any record found for a user id in database first and then save new record.
+	 * @param userId
+	 */
+	async createOtp(userId: string): Promise<string> {
 		const recentOtpRequest = await this.otpRepository.findOne({
 			where: {
-				user_id,
-				created_at: MoreThan(new Date(new Date().getTime() - this.minutesToMakeNewOtpRequest * 60 * 1000)),
+				user_id: userId,
+				created_at: MoreThan(new Date(new Date().getTime() - this.minutesToMakeNewOtpRequest)),
 			},
 		});
 		if (recentOtpRequest) {
 			throw new HttpException(
-				`Please wait ${this.minutesToMakeNewOtpRequest} minutes to make new request.`,
+				`Please wait ${this.minutesToMakeNewOtpRequest / 60000} minutes to make new request.`,
 				HttpStatus.TOO_MANY_REQUESTS,
 			);
 		}
 		const otp = generateOtp();
 		const hashedOtp = await hash(otp, env.SALT_ROUND);
 		const saveOtp = this.otpRepository.create({
-			user_id,
+			user_id: userId,
 			otp_code: hashedOtp,
-			expires_at: new Date(new Date().getTime() + this.otpExpirationMinutes * 60 * 1000),
+			expires_at: new Date(new Date().getTime() + this.otpExpirationMinutes),
 		});
-		await this.otpRepository.delete({ user_id });
+		await this.otpRepository.delete({ user_id: userId });
 		await this.otpRepository.save(saveOtp);
 		return otp;
 	}
 
-	async validateOtp(user_id: string, otp: string): Promise<boolean> {
+	/**
+	 * Checks and verifies the OTP code provided in the database and returns true if success.
+	 * This also removes the code from the database after it is verified successfully.
+	 * @param userId
+	 * @param otp
+	 */
+	async validateThenRemoveOtp(userId: string, otp: string): Promise<boolean> {
 		const validOtp = await this.otpRepository.findOne({
-			where: { user_id, expires_at: MoreThan(new Date()) },
+			where: { user_id: userId, expires_at: MoreThan(new Date()) },
 		});
 		if (!validOtp) {
 			throw new UnprocessableEntityException(errorMessage.invalidOrExpiredOtp);
@@ -52,11 +62,7 @@ export class VerificationService {
 		if (!result) {
 			throw new UnprocessableEntityException(errorMessage.invalidOrExpiredOtp);
 		}
-		if (validOtp && result) {
-			await this.otpRepository.remove(validOtp);
-			return true;
-		} else {
-			return false;
-		}
+		await this.otpRepository.remove(validOtp);
+		return true;
 	}
 }
